Update DEPLOYMENT.md / Hardware Requirements
I find it unintuitive that the deployment guide for AppFlowy Cloud suggests using AMZ EC2 or Microsoft Azure VMs, which are paid services that are hosted at Amazon / Microsoft data servers. I get that the point is that the server needs to be available / reliable and those cloud compute services are. But aren't we targeting people who want to use free opensource software here?