fix: Add postgres envvar to dev.env
The postgres envvar is missing in the dev.env, and this results in
postgres' health check failing as it uses these envvars. this makes the
health check default envvar like `POSTGRES_USER` to blank string.
For some reason it is not enough just to specify them in the environment
section of the docker compose file so add these to dev.env just like
deploy.env